+++ pkg/common/constants/constants.go
@@ -71,8 +71,6 @@ const (
 	GitLabOAuthConfigClientIdFileName       = "id"
 	GitLabOAuthConfigClientSecretFileName   = "secret"
 	OAuthScmConfiguration                   = "oauth-scm-configuration"
-	AccessToken                             = "access_token"
-	IDToken                                 = "id_token"
 
 	// Labels
 	KubernetesComponentLabelKey = "app.kubernetes.io/component"
