+++ pkg/common/operator-defaults/defaults_test.go
@@ -13,9 +13,12 @@ package defaults
 
 import (
 	"fmt"
+	"reflect"
 	"testing"
 
+	"github.com/devfile/devworkspace-operator/pkg/infrastructure"
 	chev2 "github.com/eclipse-che/che-operator/api/v2"
+	"github.com/stretchr/testify/assert"
 )
 
 func TestCorrectImageName(t *testing.T) {
@@ -82,4 +85,135 @@ func TestCorrectAirGapPatchedImage(t *testing.T) {
 
 func makeAirGapImagePath(hostname, org, nameAndTag string) string {
 	return fmt.Sprintf("%s/%s/%s", hostname, org, nameAndTag)
+}
+
+func TestIsAccesTokenConfigured(t *testing.T) {
+	t.Run("TestIsAccesTokenConfigured when access_token defined", func(t *testing.T) {
+		cheCluster := &chev2.CheCluster{
+			Spec: chev2.CheClusterSpec{
+				Networking: chev2.CheClusterSpecNetworking{
+					Auth: chev2.Auth{
+						IdentityToken: "access_token",
+					},
+				}},
+		}
+		assert.True(t, IsAccesTokenConfigured(cheCluster), "'access_token' should be activated")
+	})
+	t.Run("TestIsAccesTokenConfigured when id_token defined", func(t *testing.T) {
+		cheCluster := &chev2.CheCluster{
+			Spec: chev2.CheClusterSpec{
+				Networking: chev2.CheClusterSpecNetworking{
+					Auth: chev2.Auth{
+						IdentityToken: "id_token",
+					},
+				}},
+		}
+		assert.False(t, IsAccesTokenConfigured(cheCluster), "'access_token' should not be activated")
+	})
+}
+
+func TestGetIdentityToken(t *testing.T) {
+	t.Run("TestGetIdentityToken when access_token defined in config and k8s", func(t *testing.T) {
+		cheCluster := &chev2.CheCluster{
+			Spec: chev2.CheClusterSpec{
+				Networking: chev2.CheClusterSpecNetworking{
+					Auth: chev2.Auth{
+						IdentityToken: "access_token",
+					},
+				}},
+		}
+		infrastructure.InitializeForTesting(infrastructure.Kubernetes)
+
+		assert.Equal(t, "access_token", GetIdentityToken(cheCluster),
+			"'access_token' should be used")
+	})
+
+	t.Run("TestGetIdentityToken when id_token defined in config and k8s", func(t *testing.T) {
+		cheCluster := &chev2.CheCluster{
+			Spec: chev2.CheClusterSpec{
+				Networking: chev2.CheClusterSpecNetworking{
+					Auth: chev2.Auth{
+						IdentityToken: "id_token",
+					},
+				}},
+		}
+		infrastructure.InitializeForTesting(infrastructure.Kubernetes)
+
+		assert.Equal(t, "id_token", GetIdentityToken(cheCluster),
+			"'id_token' should be used")
+	})
+
+	t.Run("TestGetIdentityToken when no defined token in config and k8s", func(t *testing.T) {
+		cheCluster := &chev2.CheCluster{
+			Spec: chev2.CheClusterSpec{
+				Networking: chev2.CheClusterSpecNetworking{
+					Auth: chev2.Auth{},
+				}},
+		}
+		infrastructure.InitializeForTesting(infrastructure.Kubernetes)
+
+		assert.Equal(t, "id_token", GetIdentityToken(cheCluster),
+			"'id_token' should be used")
+	})
+
+	t.Run("TestGetIdentityToken when access_token defined in config and openshift", func(t *testing.T) {
+		cheCluster := &chev2.CheCluster{
+			Spec: chev2.CheClusterSpec{
+				Networking: chev2.CheClusterSpecNetworking{
+					Auth: chev2.Auth{
+						IdentityToken: "access_token",
+					},
+				}},
+		}
+		infrastructure.InitializeForTesting(infrastructure.OpenShiftv4)
+
+		assert.Equal(t, "access_token", GetIdentityToken(cheCluster),
+			"'access_token' should be used")
+	})
+
+	t.Run("TestGetIdentityToken when id_token defined in config and openshift", func(t *testing.T) {
+		cheCluster := &chev2.CheCluster{
+			Spec: chev2.CheClusterSpec{
+				Networking: chev2.CheClusterSpecNetworking{
+					Auth: chev2.Auth{
+						IdentityToken: "id_token",
+					},
+				}},
+		}
+		infrastructure.InitializeForTesting(infrastructure.OpenShiftv4)
+
+		assert.Equal(t, "id_token", GetIdentityToken(cheCluster),
+			"'id_token' should be used")
+	})
+
+	t.Run("TestGetIdentityToken when no defined token in config and openshift", func(t *testing.T) {
+		cheCluster := &chev2.CheCluster{
+			Spec: chev2.CheClusterSpec{
+				Networking: chev2.CheClusterSpecNetworking{
+					Auth: chev2.Auth{},
+				}},
+		}
+		infrastructure.InitializeForTesting(infrastructure.OpenShiftv4)
+
+		assert.Equal(t, "access_token", GetIdentityToken(cheCluster),
+			"'access_token' should be used")
+	})
+
+}
+
+func TestGetDefaultIdentityToken(t *testing.T) {
+	var tests = []struct {
+		infrastructure infrastructure.Type
+		identityToken  string
+	}{
+		{infrastructure.OpenShiftv4, "access_token"},
+		{infrastructure.Kubernetes, "id_token"},
+		{infrastructure.Unsupported, "id_token"},
+	}
+	for _, test := range tests {
+		infrastructure.InitializeForTesting(test.infrastructure)
+		if actual := GetDefaultIdentityToken(); !reflect.DeepEqual(test.identityToken, actual) {
+			t.Errorf("Test Failed. Expected '%s', but got '%s'", test.identityToken, actual)
+		}
+	}
 }
